cast size_t to clear warning

+++ chap02/time_server.c
@@ -107,13 +107,13 @@ int main() {
     printf("Sending data...\n");
     const char *response = "HTTP/1.1 200 OK\r\nConnection: close\r\nContent-Type: text/plain\r\n\r\nLocal time is: ";
     int bytes_sent = send(socket_client, response, strlen(response), 0);
-    printf("Sent %d of %lu bytes.\n", bytes_sent, strlen(response));
+    printf("Sent %d of %d bytes.\n", bytes_sent, (int)strlen(response));
 
     time_t timer;
     time(&timer);
     char *time_msg = ctime(&timer);
     bytes_sent = send(socket_client, time_msg, strlen(time_msg), 0);
-    printf("Sent %d of %lu bytes.\n", bytes_sent, strlen(time_msg));
+    printf("Sent %d of %d bytes.\n", bytes_sent, (int)strlen(time_msg));
 
 
     printf("Closing connection...\n");
